requirements.txt
huawei-lte-api==1.7.3
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3
schedule==1.2.0
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1 